requirement.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
PyPDF2==3.0.1
docx2txt==0.8
faiss-cpu==1.7.4
openai==1.12.0
python-dotenv==1.0.1
PyMuPDF==1.23.8
pandas==2.2.0
tabula-py==2.9.0
camelot-py==0.11.0
langchain-community==0.0.21
langchain-huggingface==0.0.8
sentence-transformers==2.3.1
numpy==1.26.4
Pillow==10.2.0
python-docx==1.1.0